refactor(parser): preprocess の共通 utility を preprocess/utils.ts に抽出
iftags/preprocess.ts に閉じていた `computeBracketDepths` / `maskRawRegions` /
`makeUniqueSentinels` / `restorePlaceholders` を共通モジュール
`packages/parser/src/parser/preprocess/utils.ts` に移し、iftags 側は import に
切り替えた。次コミットで追加する opener-embedded [[#if]] 等の preprocess も
同じ utility を共有する。
